ModelingTools: cast double to float to resolve clang build error #rb none #rnx #jira UE-94331
#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: ryan.schmidt
#ROBOMERGE-SOURCE: CL 13678729 via CL 13678734 via CL 13678738 via CL 13678745
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v707-13641620)

[CL 13681437 by ryan schmidt in Dev-Networking branch]

+++ Engine/Plugins/Experimental/MeshModelingToolset/Source/MeshModelingToolsEditorOnly/Private/MeshTangentsTool.cpp
@@ -318,7 +318,7 @@ void UMeshTangentsTool::OnTangentsUpdated(const TUniquePtr<FMeshTangentsd>& NewR
 					double MaxAngleDeg = FMathd::Max(TangentMikkt.AngleD(TangentNew), BitangentMikkt.AngleD(BitangentNew));
 					if (MaxAngleDeg > 0.5)
 					{
-						FMikktDeviation Deviation{ MaxAngleDeg, Index, j, Verts[j], TangentMikkt, BitangentMikkt, TangentNew, BitangentNew };
+						FMikktDeviation Deviation{ static_cast<float>(MaxAngleDeg), Index, j, Verts[j], TangentMikkt, BitangentMikkt, TangentNew, BitangentNew };
 						Deviations.Add(Deviation);
 					}
 				}
